+++ code/cgame/cg_players.c
@@ -7732,7 +7732,7 @@ void CG_ResetPlayerEntity( centity_t *cent )
 	}
 
 	if ( cg_debugPosition.integer ) {
-		CG_Printf("%i ResetPlayerEntity yaw=%i\n", cent->currentState.number, cent->pe.torso.yawAngle );
+		CG_Printf("%i ResetPlayerEntity yaw=%i\n", cent->currentState.number, (int)cent->pe.torso.yawAngle );
 	}
 }
